+++ tests/policies_tests.rs
@@ -71,3 +71,30 @@ fn test_operator_policies() -> Result<(), Box<dyn std::error::Error>> {
 
     Ok(())
 }
+
+#[test]
+fn test_policies_declare_list_and_delete() -> Result<(), Box<dyn std::error::Error>> {
+    let policy_name = "test_policies_declare_list_and_delete";
+
+    run_succeeds([
+        "policies",
+        "declare",
+        "--name",
+        policy_name,
+        "--pattern",
+        "foo-.*",
+        "--apply-to",
+        "queues",
+        "--priority",
+        "123",
+        "--definition",
+        "{\"max-length\": 20}",
+    ]);
+
+    run_succeeds(["policies", "list"])
+        .stdout(predicate::str::contains(policy_name).and(predicate::str::contains("20")));
+    run_succeeds(["policies", "delete", "--name", policy_name]);
+    run_succeeds(["policies", "list"]).stdout(predicate::str::contains(policy_name).not());
+
+    Ok(())
+}